Rip out Flow
Maybe after a few more releases this will be up to par, but for now I'm
spending too much time fighting it.

Specifically:

I discovered that it's support for exhaustiveness checking in disjoint
unions wasn't up to par, and thus was finding quite a few false
positives. They're working on it. See
https://github.com/facebook/flow/issues/451. Until that's complete I
can't really recommend it.