Initialize dashboard when it mounts
I added the beginnings of the dashboard reducer, and a higher order
component to encapsulate initialization behavior.

The higher order component is primarily to allow me to keep business
logic components in the stateless functional style. The primary drawback
I see is that the initialization function is passed as a string to call
on the incoming component props. I'd like to be more strict than that,
but it's a small price to pay I think. The drawbacks are mostly
minimized by keeping a tight locality of reference on the function name
and it's use as a string.