Automatically fetch session info when the dashboard is initialized.
A couple of points to note:

* Babel required a new plugin to handle JS spread syntax :(
* I'm using the redux-side-effect middleware to handle the asynchronous
web request. I like how it turned out, but I also dislike impure reducers.
So far though it seems worth it.
* The web request goes through a slight transformation process to
transform the data to be most convenient to the dashboard. This could
possibly have been done as an API call, but it seemed prudent to do it
here and not bother the API with specific client needs.
* I've added named PropType definitions to better communicate the data
model.

Potential improvements:

* Possibly use Ramda.invoker to call `json()` and chain promises. This
could keep me from needing the internal lambda.
* Individual Session/Answerer prop types can be abstracted.