Extract StatefulForm into it's own component
The usage of StatefulForm as a higher-order component was causing a
mixture of props that were received as props, and props that were
received from the StatefulForm wrapper. This makes the components harder
to test because it's unclear what comes from where. Extracting the
StatefulForm into a more traditional component both clarifies the
component structure, and cleans up props. The only thing that is a bit
unorthodox is passing this "form" prop that is really a function taking
the submit and change handlers for the form. Many would do this as
children of the StatefulForm component but then there's no way to get
the handlers.