Default to empty arrays
I couldn't track down the source of this problem, but sometimes
this code is reached where `user.answererSessions` and
`user.questionerSessions` are undefined. They should probably default
and/or be properly set to empty arrays.

This patch is a work-around, but I am not sure where this should
be fixed proper.